EDit

+++ app/Http/Livewire/Admin/AdminShow.php
@@ -15,11 +15,7 @@ class AdminShow extends Component
     public $add ;
     public $i;
     
-    public function removeAdmin(){
-        $this->add = "Hello From Function";
-        $record = Admin::where('id', 2)->delete();
-
-    }
+  
     
     public function render()
     {
@@ -27,13 +23,11 @@ class AdminShow extends Component
         return view('livewire.admin.admin-show');
     }
 
-    
-    public function destroy($i)
-    {
-        if ($i) {
-            $record = Admin::where('id', $i);
-            $record->delete();
-        }
+    public function deleteAdmin($id){
+        $record = Admin::where('id',$id)->delete();
+
     }
+    
+   
   
 }
